app insight/monitoring added to api + web.
