+++ .env-template
@@ -1,3 +1,3 @@
 # Copy from the "S118"-subscription in the Azure Portal. (lcm-dev/lcmdevstorage/Access keys)
 TABLE_KEY=
-APPINSIGHTS_CONSTRING=
+APPINSIGHTS_CON_STRING=
