+++ web/Dockerfile
@@ -3,6 +3,7 @@ ENV VITE_APPLICATION_OWNER=ovbra@equinor.com
 ENV VITE_CLIENT_ID=1dbc1e96-268d-41ad-894a-92a9fb85f954
 ENV VITE_SCOPES="api://lost-circulation-material-api/Optimization.All.All"
 ARG UID=1000
+ARG APPINSIGHTS_CON_STRING
 
 WORKDIR /code
 ADD package.json package-lock.json /code/
@@ -14,7 +15,7 @@ USER 1000:1000
 CMD ["npm","run", "start"]
 
 FROM base AS build
-RUN npm run build
+RUN export VITE_APPINSIGHTS_CON_STRING=$(echo $APPINSIGHTS_CON_STRING|base64 -d) && npm run build
 
 FROM node:23-alpine AS prod
 RUN npm install -g serve
