+++ web/src/applicationInsight.ts
@@ -1,7 +1,7 @@
 import { ReactPlugin } from '@microsoft/applicationinsights-react-js'
 import { ApplicationInsights } from '@microsoft/applicationinsights-web'
 
-const connectionString = import.meta.env.VITE_APPINSIGHTS_CONSTRING || ''
+const connectionString = import.meta.env.VITE_APPINSIGHTS_CON_STRING || ''
 
 export const reactPlugin = new ReactPlugin()
 export const appInsights = new ApplicationInsights({
@@ -16,5 +16,5 @@ export const appInsights = new ApplicationInsights({
 if (connectionString) {
   appInsights.loadAppInsights()
 } else {
-  console.error('Application Insights connection string is not defined');
+  console.error('Application Insights connection string is not defined')
 }
